- Refaktoryzacja skryptu mainpage.js na mniejsze pliki, by ułatwić dalszą pracę.
- Podstawy backendu do synchronizacji z dyskiem google
