+++ Backend/.env
@@ -6,3 +6,5 @@ GOOGLE_CLIENT_ID=409697200173-003cd5hk14iqmlip5irspoqk5j5vjebv.apps.googleuserco
 GOOGLE_CLIENT_SECRET=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
 FACEBOOK_APP_ID=559400323569163
 FACEBOOK_APP_SECRET=ebc2532d05f6b806305c40af1bee4182
+GOOGLE_REDIRECT_URI=http://localhost:3000/api/auth/google/callback
+GOOGLE_API_SCOPE=https://www.googleapis.com/auth/drive.file
