+++ Backend/models/Folder.js
@@ -5,7 +5,11 @@ const FolderSchema = new mongoose.Schema({
     name: { type: String, required: true },
     description: { type: String },
     parent: { type: mongoose.Schema.Types.ObjectId, ref: 'Folder', default: null },
-    createdAt: { type: Date, default: Date.now }
+    createdAt: { type: Date, default: Date.now },
+    googleDriveId: { type: String },
+    syncedFromDrive: { type: Boolean, default: false },
+    syncedToDrive: { type: Boolean, default: false },
+    lastSyncDate: { type: Date }
 });
 
 module.exports = mongoose.model('Folder', FolderSchema);
