+++ Backend/package.json
@@ -10,6 +10,7 @@
     "exiftool-vendored": "^29.3.0",
     "express": "^4.18.2",
     "google-auth-library": "^9.15.1",
+    "googleapis": "^148.0.0",
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.0.3",
     "multer": "^1.4.5-lts.2"
